move PreparedStatementsTests
